twosumOn generalized to Foldable
The mildly annoying thing here is that I had it working with a plain
foldl', but that does require always looking through the entire
structure to get the answer.  Using toList seems a little less
elegant, but it means I only need to consume as much as necessary to
find an answer.